Changed virtual exchange naming to use semicolon instead of underscore
The underscore had a dual-usage between the config file (to represent
spaces) and in the code (to represent a virtual exchange).  This
causes a problem when the config specifies an exchange and uses an
underscore followed by four or more characters.  The code would treat
this configured exchange as a virtual one and end up looking for the
configuration of another exchange.

No changes are required on the user's part.  A simple `update` call
will fix the naming convention in the user's files.